custom pagination from products as global
