--- project_1444/product/pagination.py
@@ -1,21 +1,0 @@
-# product/pagination.py
-from rest_framework.response import Response
-from rest_framework.pagination import PageNumberPagination
-
-
-class Pagination(PageNumberPagination):
-    page_size = 10
-    page_size_query_param = "page_size"
-    max_page_size = 100
-
-    def get_paginated_response(self, data):
-        return Response(
-            {
-                "links": {
-                    "next": self.get_next_link(),
-                    "previous": self.get_previous_link(),
-                },
-                "count": self.page.paginator.count,
-                "results": data,
-            }
-        )
